requirements.txt
click
fastapi
uvicorn
sqlalchemy
python-dotenv
openai
jinja2
python-multipart
requests
ffmpeg-python
inputimeout
elevenlabs
websockets 
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
requests
python-dotenv
instructor
pydantic 
google-search-results
jinja2
beautifulsoup4
scipy
# Add parallax-maker from GitHub
parallax-maker @ git+https://github.com/provos/parallax-maker.git 